Ajout petite page front et nouvelle méthode pour les frequences et la grid_build

+++ server/dictionnary/src/include/grid_build.h
@@ -4,6 +4,6 @@
 struct Letter_occurence
 {
     char letter[2];
-    int count;
+    float count;
 };
 typedef struct Letter_occurence Letter_occurence;
